+++ EduVPN/Networking/DiscoveryDataFetcher.swift
@@ -10,17 +10,42 @@ import Moya
 import PromiseKit
 
 enum DiscoveryDataFetcherError: Error {
-    case dataCouldNotBeVerified
+    case dataCouldNotBeVerified(url: URL)
     case dataNotFoundInCache
     case dataNotFoundInAppBundle
+    case versionNumberNotFound(url: URL)
+    case versionNumberDecreased(url: URL, previousVersion: Int, newVersion: Int)
 }
 
 extension DiscoveryDataFetcherError: AppError {
     var summary: String {
         switch self {
-        case .dataCouldNotBeVerified: return "Discovery data could not be verified"
-        case .dataNotFoundInCache: return "Discovery data not found in cache"
-        case .dataNotFoundInAppBundle: return "Discovery data not found in app bundle"
+        case .dataCouldNotBeVerified:
+            return "Discovery data could not be verified"
+        case .dataNotFoundInCache:
+            return "Discovery data not found in cache"
+        case .dataNotFoundInAppBundle:
+            return "Discovery data not found in app bundle"
+        case .versionNumberNotFound:
+            return "Discovery data doesn't contain version number"
+        case .versionNumberDecreased:
+            return "Discovery data was not updated because the version number has decreased"
+        }
+    }
+    var detail: String {
+        switch self {
+        case .dataCouldNotBeVerified(let url):
+            return "URL: \(url)"
+        case .dataNotFoundInCache, .dataNotFoundInAppBundle:
+            return ""
+        case .versionNumberNotFound(let url):
+            return "URL: \(url)"
+        case .versionNumberDecreased(let url, let previousVersion, let newVersion):
+            return """
+            URL: \(url)
+            Previous version: \(previousVersion)
+            New version: \(newVersion)
+            """
         }
     }
 }
@@ -38,6 +63,20 @@ struct DiscoveryDataFetcher {
         case appBundle // Get the data from the JSON file contained in the app bundle
         case cache // Get the data from the web cache
         case server // Download the data afresh from the discovery server
+    }
+
+    struct CacheContents {
+        let dataURL: URL
+        let dataResponse: CachedURLResponse
+        let signatureURL: URL
+        let signatureResponse: CachedURLResponse
+
+        func store(to cache: URLCache) throws {
+            for (url, response) in [(dataURL, dataResponse), (signatureURL, signatureResponse)] {
+                let request = try URLRequest(url: url, method: .get)
+                cache.storeCachedResponse(response, for: request)
+            }
+        }
     }
 
     static var diskCache: URLCache {
@@ -59,47 +98,154 @@ struct DiscoveryDataFetcher {
         switch origin {
         case .appBundle:
             return Promise { seal in
-                let splitLastPathComponent = dataURL.lastPathComponent.split(separator: ".")
-                guard splitLastPathComponent.count == 2 else {
-                    seal.reject(DiscoveryDataFetcherError.dataNotFoundInAppBundle)
-                    return
-                }
-                let fileName = String(splitLastPathComponent[0])
-                let fileExtension = String(splitLastPathComponent[1])
-                guard let includedServerListURL = Bundle.main.url(forResource: fileName, withExtension: fileExtension) else {
-                    seal.reject(DiscoveryDataFetcherError.dataNotFoundInAppBundle)
-                    return
-                }
-                let data = try Data(contentsOf: includedServerListURL)
+                let data = try getFromAppBundle(dataURL: dataURL)
                 seal.fulfill(data)
             }
         case .cache:
             return Promise { seal in
-                guard let dataResponse = diskCache.cachedResponse(for: URLRequest(url: dataURL)),
-                    let signatureResponse = diskCache.cachedResponse(for: URLRequest(url: signatureURL)) else {
-                        seal.reject(DiscoveryDataFetcherError.dataNotFoundInCache)
-                        return
-                }
-                seal.fulfill(try verify(data: dataResponse.data,
-                                        signature: signatureResponse.data,
-                                        publicKeys: publicKeys))
+                let (data, _) = try getFromDiskCache(
+                    dataURL: dataURL, signatureURL: signatureURL,
+                    publicKeys: publicKeys)
+                seal.fulfill(data)
             }
         case .server:
-            let dataProvider = MoyaProvider<Target>(session: cachedSession)
-            let dataPromise = dataProvider.request(target: Target(dataURL))
-
-            let signatureProvider = MoyaProvider<Target>(session: cachedSession)
-            let signaturePromise = signatureProvider.request(target: Target(signatureURL))
-
-            return when(fulfilled: dataPromise, signaturePromise)
-                .map { dataResponse, signatureResponse in
-                    return try verify(data: dataResponse.data, signature: signatureResponse.data,
-                                      publicKeys: publicKeys)
-                }
+            return getFromRemoteServerWithRollbackPrevention(
+                dataURL: dataURL,
+                signatureURL: signatureURL,
+                publicKeys: publicKeys)
         }
     }
 
-    private static func verify(data: Data, signature: Data, publicKeys: [Data]) throws -> Data {
+    private static func getFromAppBundle(dataURL: URL) throws -> Data {
+        let splitLastPathComponent = dataURL.lastPathComponent.split(separator: ".")
+        guard splitLastPathComponent.count == 2 else {
+            throw DiscoveryDataFetcherError.dataNotFoundInAppBundle
+        }
+        let fileName = String(splitLastPathComponent[0])
+        let fileExtension = String(splitLastPathComponent[1])
+        guard let includedServerListURL = Bundle.main.url(forResource: fileName, withExtension: fileExtension) else {
+            throw DiscoveryDataFetcherError.dataNotFoundInAppBundle
+        }
+        return try Data(contentsOf: includedServerListURL)
+    }
+
+    private static func getFromDiskCache(dataURL: URL, signatureURL: URL, publicKeys: [Data]) throws -> (Data, CacheContents) {
+        guard let dataResponse = diskCache.cachedResponse(for: URLRequest(url: dataURL)),
+              let signatureResponse = diskCache.cachedResponse(for: URLRequest(url: signatureURL)) else {
+            throw DiscoveryDataFetcherError.dataNotFoundInCache
+        }
+        let data = try verify(data: dataResponse.data,
+                          signature: signatureResponse.data,
+                          publicKeys: publicKeys,
+                          dataURL: dataURL)
+        let cacheContents = CacheContents(
+            dataURL: dataURL,
+            dataResponse: dataResponse,
+            signatureURL: signatureURL,
+            signatureResponse: signatureResponse)
+        return (data, cacheContents)
+    }
+
+    private static func getFromRemoteServer(dataURL: URL, signatureURL: URL, publicKeys: [Data]) -> Promise<Data> {
+        let dataProvider = MoyaProvider<Target>(session: cachedSession)
+        let dataPromise = dataProvider.request(target: Target(dataURL))
+
+        let signatureProvider = MoyaProvider<Target>(session: cachedSession)
+        let signaturePromise = signatureProvider.request(target: Target(signatureURL))
+
+        return when(fulfilled: dataPromise, signaturePromise)
+            .map { dataResponse, signatureResponse in
+                return try verify(data: dataResponse.data, signature: signatureResponse.data,
+                                  publicKeys: publicKeys, dataURL: dataURL)
+            }
+    }
+
+    private static func getFromRemoteServerWithRollbackPrevention(
+        dataURL: URL, signatureURL: URL, publicKeys: [Data]) -> Promise<Data> {
+        let (previousData, previousVersion, cacheContents): (Data?, Int?, CacheContents?) = {
+            if let (data, cacheContents) = try? getFromDiskCache(
+                dataURL: dataURL, signatureURL: signatureURL,
+                publicKeys: publicKeys) {
+                let version = versionValue(from: data)
+                return (data, version, cacheContents)
+            } else if let data = try? getFromAppBundle(dataURL: dataURL) {
+                let version = versionValue(from: data)
+                return (data, version, nil)
+            } else {
+                return (nil, nil, nil)
+            }
+        }()
+        return firstly {
+            getFromRemoteServer(
+                dataURL: dataURL, signatureURL: signatureURL,
+                publicKeys: publicKeys)
+        }.map { data in
+            if previousData == data {
+                // Data hasn't changed
+                return data
+            }
+            if let newVersion = versionValue(from: data) {
+                if let previousVersion = previousVersion {
+                    if newVersion > previousVersion {
+                        // Data has changed and version number has gone up
+                        return data
+                    } else if newVersion < previousVersion {
+                        // Version number has gone down. We show an error.
+                        try? cacheContents?.store(to: diskCache)
+                        throw DiscoveryDataFetcherError.versionNumberDecreased(
+                            url: dataURL,
+                            previousVersion: previousVersion,
+                            newVersion: newVersion)
+                    } else {
+                        // Version number has not changed, so we use the old version.
+                        // No error to be shown.
+                        try? cacheContents?.store(to: diskCache)
+                        return data
+                    }
+                } else {
+                    // No previous version. Can't really happen.
+                    return data
+                }
+            } else {
+                throw DiscoveryDataFetcherError.versionNumberNotFound(url: dataURL)
+            }
+        }
+    }
+
+    private static func versionValue(from data: Data) -> Int? {
+        versionValueUsingRegex(from: data) ?? versionValueUsingParser(from: data)
+    }
+
+    private static func versionValueUsingRegex(from data: Data) -> Int? {
+        let upperBound = min(100, data.count)
+        // swiftlint:disable:next force_unwrapping
+        guard let commaIndex = data[0 ..< upperBound].firstIndex(of: Character(",").asciiValue!) else {
+            return nil
+        }
+        guard let string = String(data: data[0 ..< commaIndex], encoding: .utf8) else {
+            return nil
+        }
+        guard let regex = try? NSRegularExpression(pattern: "^\\s*\\{\\s*\"v\"\\s*:\\s*(\\d+)", options: []) else {
+            return nil
+        }
+        guard let result = regex.firstMatch(in: string, options: [], range: NSRange(location: 0, length: string.count)) else {
+            return nil
+        }
+        let versionRange = result.range(at: 1)
+        let versionStart = string.utf16.index(string.utf16.startIndex, offsetBy: versionRange.lowerBound)
+        let versionEnd = string.utf16.index(string.utf16.startIndex, offsetBy: versionRange.upperBound)
+        let versionString = string[versionStart ..< versionEnd]
+        return Int(versionString)
+    }
+
+    private static func versionValueUsingParser(from data: Data) -> Int? {
+        guard let versionable = try? JSONDecoder().decode(VersionableDiscoveryData.self, from: data) else {
+            return nil
+        }
+        return versionable.version
+    }
+
+    private static func verify(data: Data, signature: Data, publicKeys: [Data], dataURL: URL) throws -> Data {
         let signature = try SignatureHelper.minisignSignatureFromFile(data: signature)
         for publicKey in publicKeys {
             let isValid = SignatureHelper.isSignatureValid(
@@ -109,6 +255,14 @@ struct DiscoveryDataFetcher {
                 return data
             }
         }
-        throw DiscoveryDataFetcherError.dataCouldNotBeVerified
+        throw DiscoveryDataFetcherError.dataCouldNotBeVerified(url: dataURL)
+    }
+}
+
+private struct VersionableDiscoveryData: Decodable {
+    let version: Int
+
+    enum CodingKeys: String, CodingKey {
+        case version = "v"
     }
 }